\name{NL_6min_2013}
\docType{data}
\alias{NL_6min_2013}
\title{New London water levels, 2013}
\description{
   Water levels from NOAA-COOPS station number 8461490 in New London, CT. Data recorded at 6-minute intervals, in meters relative to MHW, and in the GMT time zone.
}
\usage{
data(NL_6min_2013)
}
\format{
A dataframe with 87591 rows and 3 variables.
}
\source{
downloaded from National Atmospheric and Oceanic Administration's Center for Operational Oceanographic Products and Services using \code{VulnToolkit::noaa()} 
}

