\name{VulnToolkit-package}
\alias{VulnToolkit-package}
\alias{VulnToolkit}
\docType{package}
\title{A set of web scrapers and analytical tools to aid analysis of coastal wetlands and their vulnerability to sea-level rise.
}
\description{VulnToolkit includes analytical tools to download data from NOAA-COOPS stations, including tide data, harmonic constituents, and station datums. VulnToolkit also includes tools for measuring hydroperiod and assessing vulnerability to hydrologic stress.
}
\details{
\tabular{ll}{
Package: \tab VulnToolkit\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2014-04-09\cr
License: \tab MIT\cr
}

}
\author{Troy D. Hill <Hill.Troy@gmail.com>, Shimon C. Anisfeld

Maintainer: Troy D. Hill <Hill.Troy@gmail.com>

}
\references{ in preparation
}
\keyword{ package }
\seealso{
}
\examples{
}
