% Generated by roxygen2 (4.0.2): do not edit by hand
\name{harcon}
\alias{harcon}
\title{Scrapes harmonic constituent data from NOAA CO-OPS website}
\usage{
harcon(station)
}
\arguments{
\item{station}{station ID number or vector of IDs, available on CO-OPS
website (\url{http://co-ops.nos.noaa.gov/stations.html?type=Water+Levels}) or
from \code{\link{noaa.stations}}.}
}
\value{
a dataframe of harmonic constituents and their associated phases,
amplitudes, and speeds.
}
\description{
Scrapes harmonic constituent data from NOAA CO-OPS
website. Requires an internet connection.
}
\examples{
\dontrun{
bport.cons <- harcon(8467150) # Bridgeport, CT
bport.cons
}
}
\references{
\url{http://wetlandsandr.wordpress.com/}
}
\seealso{
\code{\link{noaa.stations}}
}

