\name{NL_6min_2013}
\docType{data}
\alias{NL_6min_2013}
\title{New London water levels, 2013}
\description{
   Water levels from NOAA-COOPS station #8461490 in New London, CT. Data recorded at 6-minute intervals, in meters relative to MHW, and in the GMT time zone.
}
\usage{
data(NL_6min_2013)
}
\format{
A dataframe with 87591 rows and 3 variables.
}
\source{
downloaded from http://co-ops.nos.noaa.gov/inventory.html?id=8461490 using \code{VulnToolkit::noaa()} 
}
\references{
   Center for Operational Oceanographic Products and Services 
}
